Login, Logout and Signup functionality is now fully implemented.

+++ frontend/src/components/PaletteDetails.js
@@ -1,16 +1,25 @@
 import { Delete } from '@material-ui/icons';
 import React from 'react'
 import {usePalettesContext} from '../hooks/usePalettesContext'
+import {useAuthContext} from '../hooks/useAuthContext';
 
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 
 const PaletteDetails = ({palette}) => {
   const {dispatch} = usePalettesContext()
+  const {user} = useAuthContext()
+
 
   const handleClick = async() => {
+    if(!user){
+      return
+    }
     const response = await fetch('/api/palettes/' + palette._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
     const json = await response.json()
